notes_backend/requirements.txt
Django==5.2.4
gunicorn
# psycopg2-binary
# djangorestframework
# python-decouple
# django-cors-headers


djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg2-binary
python-decouple
